@@ -1,17 +1,24 @@
 plugins {
-  id 'java'
-  id 'maven'
-  id 'com.github.ben-manes.versions' version '0.28.0'
+  id 'java-library'
+  id 'com.github.ben-manes.versions' version '0.46.0'
 }
 
 group = 'net.reini'
 version = '0.0.1-SNAPSHOT'
 
-sourceCompatibility = 9
-targetCompatibility = 9
+java {
+  sourceCompatibility = JavaVersion.VERSION_17
+}
 
 tasks.withType(JavaCompile) {
   options.encoding = 'UTF-8'
+}
+
+test {
+  useJUnitPlatform()
+  testLogging {
+    events 'skipped'
+  }
 }
 
 jar {
@@ -19,12 +26,16 @@ jar {
 }
 
 repositories {
-  jcenter()
-  mavenLocal()
+  mavenCentral()
 }
 
 dependencies {
-  testCompile 'junit:junit:4.13'
+  testImplementation 'org.assertj:assertj-core:3.24.2'
+  testImplementation 'org.junit.jupiter:junit-jupiter-api:5.9.2'
+  testImplementation 'org.mockito:mockito-junit-jupiter:5.2.0'
+
+  testRuntimeOnly 'org.junit.platform:junit-platform-runner:1.9.2'
+  testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.9.2'
 }
 
 dependencyUpdates.resolutionStrategy {
