@@ -1,17 +1,17 @@
 package card;
 
-import static org.junit.Assert.assertArrayEquals;
-import static org.junit.Assert.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertArrayEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import org.junit.Before;
-import org.junit.Test;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 
-public class MainTest {
+class MainTest {
     private byte[] data;
     private String dataString;
 
-    @Before
-    public void setUp() {
+    @BeforeEach
+    void setUp() {
         data = new byte[20];
         for (int idx=0; idx<20; idx++) {
             data[idx] = (byte) idx;
@@ -20,12 +20,12 @@ public class MainTest {
     }
 
     @Test
-    public void printHexBinary() {
+    void printHexBinary() {
         assertEquals(dataString, Main.printHexBinary(data));
     }
 
     @Test
-    public void parseHexBinary() {
+    void parseHexBinary() {
         assertArrayEquals(data, Main.parseHexBinary(dataString));
     }
 }
