@@ -1,16 +1,17 @@
 package card;
 
+import static java.nio.file.Files.newInputStream;
+import static java.nio.file.Files.newOutputStream;
+
 import java.awt.Component;
-import java.io.*;
+import java.io.File;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.OutputStream;
 import java.nio.file.Paths;
-import java.util.stream.IntStream;
 
-import javax.swing.Icon;
 import javax.swing.JFileChooser;
-import javax.swing.JOptionPane;
 import javax.swing.filechooser.FileNameExtensionFilter;
-
-import static java.nio.file.Files.*;
 
 public class StoryMania {
     public static void main(String[] args) throws IOException {
